+++ src/ui/layouts/runs/RunsTable/HeaderCols/index.tsx
@@ -26,6 +26,7 @@ import { useHistory, useSelector } from '../../../../hooks';
 import { routePaths } from '../../../../../routes/routePaths';
 import { workspaceSelectors } from '../../../../../redux/selectors';
 import { Run } from '../../../../../api/types';
+import { CustomToolTip } from '../../../common/CustomToolTip';
 
 const HeaderText = ({ text, margin }: { text: string; margin?: string }) => (
   <Paragraph
@@ -46,20 +47,6 @@ const innerBoxStyleEnable = {
   backgroundColor: '#f0ebfc',
   justifyContent: 'center',
   alignItems: 'center',
-};
-
-const customToolTip = {
-  border: '2px solid #f0ebfc',
-  borderRadius: '5px',
-  display: 'flex',
-  padding: 16,
-  // justifyContent: 'center',
-  // alignItems: 'center',
-  zIndex: 1000,
-  backgroundColor: 'white',
-  position: 'absolute',
-
-  marginBottom: '200px',
 };
 
 const innerBoxStyleDisable = {
@@ -392,39 +379,7 @@ export const useHeaderCols = ({
                 <>
                   {showToolTip === run?.id &&
                     selectedSubModelId === run?.body?.pipeline?.id && (
-                      <Box
-                        style={customToolTip as any}
-                        // style={innerBoxStyleDisable}
-                      >
-                        <div
-                          style={{
-                            width: '380px',
-                          }}
-                        >
-                          <p
-                            style={{
-                              fontSize: 18,
-                              fontWeight: 'bold',
-                              fontFamily: 'Rubik',
-                            }}
-                          >
-                            You Don't have access to this Pipeline.
-                          </p>
-                          <p
-                            style={{
-                              fontSize: 16,
-                              color: '#666c78',
-                              fontFamily: 'Rubik',
-                            }}
-                          >
-                            Please contact your admin for further information
-                            <br />
-                            or to request access.
-                            <br />
-                            {`( ${run?.body?.pipeline?.name} )`}
-                          </p>
-                        </div>
-                      </Box>
+                      <CustomToolTip name={run?.body?.pipeline?.name} />
 
                       //   <Paragraph>You Don't have acces to this Stack.</Paragraph>
                       //   <Paragraph>
@@ -558,46 +513,7 @@ export const useHeaderCols = ({
                 <>
                   {showToolTip === run?.id &&
                     selectedSubModelId === run?.body?.stack?.id && (
-                      <Box
-                        style={customToolTip as any}
-                        // style={innerBoxStyleDisable}
-                      >
-                        <div
-                          style={{
-                            width: '380px',
-                          }}
-                        >
-                          <p
-                            style={{
-                              fontSize: 18,
-                              fontWeight: 'bold',
-                              fontFamily: 'Rubik',
-                            }}
-                          >
-                            You Don't have access to this Stack.
-                          </p>
-                          <p
-                            style={{
-                              fontSize: 16,
-                              color: '#666c78',
-                              fontFamily: 'Rubik',
-                            }}
-                          >
-                            Please contact your admin for further information
-                            <br />
-                            or to request access.
-                            <br />
-                            {`( ${run?.body?.stack?.name} )`}
-                          </p>
-                        </div>
-                      </Box>
-
-                      //   <Paragraph>You Don't have acces to this Stack.</Paragraph>
-                      //   <Paragraph>
-                      //     Please, contact with you Admin for further information
-                      //     or to request access.
-                      //   </Paragraph>
-                      // </Box>
+                      <CustomToolTip name={run?.body?.stack?.name} />
                     )}
 
                   <Box style={innerBoxStyleDisable}>
